fix: fix some build errors and unused variables
